refactor some class and calls to commonsKmm
