+++ app/src/main/java/org/dhis2/usescases/eventsWithoutRegistration/eventCapture/EventCaptureModule.kt
@@ -17,11 +17,11 @@ import org.dhis2.commons.schedulers.SchedulerProvider
 import org.dhis2.data.dhislogic.DhisEnrollmentUtils
 import org.dhis2.data.forms.dataentry.SearchTEIRepository
 import org.dhis2.data.forms.dataentry.SearchTEIRepositoryImpl
-import org.dhis2.form.data.FileController
 import org.dhis2.form.data.FormValueStore
 import org.dhis2.form.data.UniqueAttributeController
 import org.dhis2.form.model.RowAction
 import org.dhis2.form.ui.FieldViewModelFactory
+import org.dhis2.mobile.commons.files.FileController
 import org.dhis2.mobileProgramRules.EvaluationType
 import org.dhis2.mobileProgramRules.RuleEngineHelper
 import org.dhis2.mobileProgramRules.RulesRepository
