+++ commons/build.gradle.kts
@@ -65,6 +65,7 @@ kotlin {
 dependencies {
     implementation(fileTree(mapOf("dir" to "libs", "include" to listOf("*.jar"))))
     api(project(":ui-components"))
+    implementation(project(":commonskmm"))
 
     api(libs.dhis2.android.sdk) {
         exclude("org.hisp.dhis", "core-rules")
